Auto-merge PR instead of pushing directly to main
Because main is a protected branch, GH Actions can't push to it and
working around this is tedious. We can use the GitHub CLI to automate
the PR merge process for us. The drawback is that we have to wait for
the required checks to pass, and we have to also approve the PR.